pyproject.toml
[project]
name = "webster"
dynamic = ["version"]
description = "no :^)"
authors = [
    {name = "hex benjamin", email = "hex@hexbenjam.in"},
]
dependencies = [
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
    "rich>=13.6.0",
    "click>=8.1.7",
    "chromadb>=0.4.15",
    "openai>=0.28.1",
    "langchain-g4f @ git+https://github.com/MIDORIBIN/langchain-gpt4free.git",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10,<3.13"
readme = "README.md"
license = {text = "MIT"}
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.version]
source = "file"
path = "src\\webster\\__init__.py"

[tool.pdm.dev-dependencies]
dev = [
    "sourcery>=1.12.0",
    "rich>=13.6.0",
]

[tool.pdm.scripts]
webster = {call = "webster.cli:start"}

[tool.pdm.build.wheel-data]
scripts = ["src\\webster\\cli.py"]